makefile: use correct clock frequency
The primary clock frequency on Fomu is 48 MHz, and this design is a
single clock domain. As a result, we don't need to use a Python script
to define various clock domain frequencies, we can simply pass it on the
command line with `--freq`.

If this design were split into multiple clock domains, we would need to
use a Python script to assign frequencies to various domains, or use a
newer version of `nextpnr` that supports assigning clock domains in a
separate file.

Signed-off-by: Sean Cross <sean@xobs.io>