Start of VECTOR and DISPLAY LIST
Start of the implementation of the vector drawer, partially working, some lines missed from the drawing.

Start of the implementation of display lists. Not integrated into the system at present.

Fix to the audio system for @sylefeb

Fix to the line drawing system, now seems to finally cope with all 8 quadrants.