Split bram blocks in vectors and display lists
Split the brams into individual attributes for the vectors and the display lists. Makes the code easier to read.

Need to find bottlenecks to improve timings.

Vectors do not always fully draw. Looks like some lines are not correctly sent to the GPU.

Display list does not produce any output.