Optimise GPU Filled Triangle
Optimise filled triangle drawing. When leaving the triangle on a line, decide if closer to the left or right, and start the walking from that side. Cuts down on the number of considered pixels outside of the triangle.

ULX3S build posted has broken audio. New build will be built, eventually, as yosys still taking forever.