PAWSLIBRARY WORK
Abstract all I/O and DISPLAY functions, so no direct hardware register access is needed.

Document the functions in the PAWSLIBRARY.

Provide functions to set bitmaps for the sprites, tilemap and blitter, update asteroids and bios to use.

GPU functions now set parameters and wait for GPU before completing, speeds up slightly the dispatch of commands to the GPU

Blitter can now double in size the tiles to 32 x 32, plus correct an error to ensure all pixels in the blitter tile are copied to the bitmap.