SMT fixes / changes /tests
Try increasing SMT instruction cache to 32 entries

Double timers to allow a set for the main and smt threads

Provide a more comprehensive demo of SMT, change asteroids to use a second thread to move the asteroids. Sprite layer - provide an alternate set of registers to allow the smt thread to control the sprites separately from the main thread - used in asteroids-smt/move_asteroids

Add in a 3 variation of the display layer ordering, bitmap just below the character map and above sprites/tilemap

Change BIOS to a single monolithic program, and use -fwhole-program to compile. Decreases output size considerably. Allows moving of SMT stack to 0x2000.

Modify library to allow for second set of timers and the alternate sprite registers.