Partially rewire in cmix cmov instructions
Start setting up for ternary bit extension instructions, cmix and cmov, not yet fully wired in. Decide if to do funnel shifts, FSL FSR[I]. Re-implement CRC32?

Wire back in ps2 port. Not working either.

Add colour to invaders and a background. Compiled with -O3 is much smoother.

Create scripts for compiling with clang or with gcc and -O3.