GPU command queue
Add a 64 entry command queue onto the GPU. Allows the CPU to calculate the next command and send to the GPU whilst the previous commands are being executed. Used, as depending on the size of the object being drawn the CPU can be waiting for the GPU for a considerable time.