Add in a pixelblock
A function for the GPU that outputs a stream of 7bit ARRGGBB or 24bit RGB pixels into a rectangle.

The pixels are streamed by a function in the PAWSlibrary that reads a set number of pixels and sends them one by one for display. This is significantly quicker than reading a pixel, and calling the GPU to plot a single pixel.

24bit RGB pixels are scaled to the PAWS colour space. For 7 bit, a colour can be defined that is not sent to the bitmap, for mask purposes.