PS/2 KEYBOARD
Finish mapping all of the basic keys on the UK keyboard, all done with the exception of PRTSCR/SYSRQ, SCROLL LOCK and PAUSE BREAK.

Basic SHIFT/CTRL/CAPSLOCK/NUMLOCK operate. ALTGR (RIGHT ALT) allows bar on key next to 1 on the UK keyboard.

Allow PS/2 keyboard to work as a joystick, using arrow keys, or numeric keypad (allows 8 directions), along with CTRL as fire 1, LEFT WINDOWS as fire 2, and LEFT ALT, RIGHT ALT, RIGHT WINDOWS, MENU AND RIGHT CTRL as extra fire keys.

PS/2 keyboard as a joystick can  be read the same as the ULX3S buttons, and the matching buttons are mapped equivalently, with no alteration to software. The extra keys are mapped to the upper 8 bits of the buttons register.

Move PS/2 and UART controllers to the 25MHz clock domain. Use a latch to control the update of the buffers from the CPU to cross the clock domain.

Add in the ability to switch the keyboard from joystick(default) to keyboard operation. BIOS sets the keyboard to joystick by default. Add in a function in the PAWSlibrary to switch the keyboard mode. Update the keyboard etst software to test all of the features of the keyboard.

SDRAM controller, ensure 26 bit addressing is being used throughout the sdram controller.

Minor change to multiplex display to allow a mode whereby the character map isn't the top layer to allow the keyboard test software to use a sprite as a cursor. Will program a hardware cursor shortly directly into the character map layer.