Documentation Improvements
Start to update the documentation to include using the keyboard as a joystick.

Enable verilator to build by wrapping some I/O parts with simulation flags.

Use a byteaccess flag in the BRAM and CACHE controllers, to make more system neutral, maps the Risc-V function3 as appropriate.

Fix a bug in the libPAWS curses emulation that fails to clear the last line when scrolling.