Optimisations and fixes
Simplify the base integer ALU, combining register-immediate and register-register optimisations. Remove start/busy flags.

Split out the M extension ALU from the base integer ALU, and split into multiplcation and divide. Remove start/busy flags from multiplication.

Add in a cropping rectangle to the bitmap writer, and allow setting of the cropping rectangle. Add functions to libPAWS, and set default to the whole screen. Update the BIOS to reset the cropping rectangle.

Update the programming guide, documenting the changes for rotation, reflection and cropping for the gpu functions. Add in the ability to print vertically using the GPU.

Remove a few not needed ++: from the GPU

Recompile software to take into account changes in the library for the rotation and reflection functions.