Expand coordinates for GPU to 16bit
Previously coordinates for the GPU were 10 bit, +/- 511, this could cause glitches when coordinates were overflowing, expanding to +/-32767 will cover most issues. Tested with outrun and does fix glitches at the bottom right of the screen where coordinates for the shoulder markings were going out of range.

Few minor optimisations in float32, removing a couple of switches that can be replaced with if else.

Update float16 to same standard as float32, tested as best I can at present.

GPU crop, limit to screen when setting cropping rectangle.

Outrun, follow the next part of the guide and add in tunnels.