FOMU and GPU
FOMU - resurrect j1eforth for the FOMU, tested on my HACKER board, have build for PVT but cannot test.

Various optimisations to GPU clock domain, simplifying in bitmapwriter, and in GPU.

Rename parameters in the blitters to tile, scale and action, simplify preparation for drawing rectangles,   and simplification to the blitters, removing excess variables.