Optimisations
Use a 2 x and a 3 x register file for integer and floating-point, remove 3rd register from integer path.

Use if instead of ternary in floatcompare in float32

Use if instead of switch for ps2data_ext in ps2

Move back to state machine for main CPU loop, and improve register writing.

Use a single csr for cycletime and instret, plus use a table for the fcsr register.

Use if instead of switch when detecting NAN in floatminmax in the FPU.

Continue work on fat32 file browser in preparation for porting to the BIOS, does not yet enter directories.